Extract Item.Create
